Environment Variable Substitution
This patch enables a feature called Environment Variable Substitution.
If a value returned by the GetString or GetStringSlice functions
contains an environment variable in the form of $KEY, the $KEY
placeholder will be replaced with the environment variable's actual
value. Matching is performed against a list of current environment
variable keys only, which means the value '$NOT_DEFINED' may never be
replaced if there is no environment variable defined as `NOT_DEFINED`.

It is possible to disable this feature using the global flag,
DisableEnvVarSubstitution. Setting the environment variable
`GOFIG_DISABLE_ENVVAR_SUBSTITUTION` to true also disables this feature.